gatsby methods clean up
