+++ packages/gatsby-mdx/gatsby-node.js
@@ -1,120 +1,26 @@
 const path = require("path");
-const fs = require("fs-extra");
-const merge = require("lodash/merge");
-const escapeStringRegexp = require("escape-string-regexp");
 const defaultOptions = require("./utils/default-options");
-const extractExports = require("./utils/extract-exports");
 const mdx = require("./utils/mdx");
 
 /**
  * Create Mdx nodes from MDX files.
  */
-exports.onCreateNode = require("./on-create-node");
+exports.onCreateNode = require("./gatsby/on-create-node");
 
 /**
  * Add additional fields to MDX nodes
  */
-exports.setFieldsOnGraphQLNodeType = require("./extend-node-type");
+exports.setFieldsOnGraphQLNodeType = require("./gatsby/extend-node-type");
 
 /**
  * Add frontmatter as page context for MDX pages
  */
-exports.onCreatePage = async ({ page, actions }, pluginOptions) => {
-  const { createPage, deletePage } = actions;
-  const { extensions, ...options } = defaultOptions(pluginOptions);
-  const ext = path.extname(page.component);
-
-  if (extensions.includes(ext)) {
-    const content = await fs.readFile(page.component, "utf8");
-    const code = await mdx(content, options);
-
-    // grab the exported frontmatter
-    const { frontmatter } = extractExports(code);
-
-    deletePage(page);
-    createPage(
-      merge(
-        {
-          context: {
-            frontmatter: {
-              ...frontmatter
-            }
-          }
-        },
-        page
-      )
-    );
-  }
-};
+exports.onCreatePage = require("./gatsby/on-create-page");
 
 /**
  * Add the webpack config for loading MDX files
  */
-exports.onCreateWebpackConfig = (
-  { stage, loaders, plugins, actions, getNodes },
-  pluginOptions
-) => {
-  const options = defaultOptions(pluginOptions);
-  const testPattern = new RegExp(
-    options.extensions.map(ext => `${escapeStringRegexp(ext)}$`).join("|")
-  );
-  const mdxTestPattern = new RegExp(
-    options.extensions
-      .concat(".deck-mdx")
-      .map(ext => `${escapeStringRegexp(ext)}$`)
-      .join("|")
-  );
-
-  const decks = options.decks.map(ext => `${escapeStringRegexp(ext)}`);
-
-  actions.setWebpackConfig({
-    module: {
-      rules: [
-        {
-          test: /\.js$/,
-          include: path.resolve(__dirname, ".cache/gatsby-mdx"),
-          use: [loaders.js()]
-        },
-        {
-          test: testPattern,
-          exclude: decks,
-          use: [
-            loaders.js(),
-            {
-              loader: "gatsby-mdx/mdx-loader",
-              options: {
-                getNodes,
-                pluginOptions: options
-              }
-            }
-          ]
-        },
-        {
-          test: mdxTestPattern,
-          include: decks,
-          use: [
-            loaders.js(),
-            { loader: "gatsby-mdx/mdx-deck-post-loader" },
-            { loader: "mdx-deck/loader" }
-          ]
-        },
-        {
-          test: /.deck-mdx$/,
-          use: [
-            loaders.js(),
-            { loader: "gatsby-mdx/mdx-deck-post-loader" },
-            { loader: "mdx-deck/loader" }
-          ]
-        }
-      ]
-    },
-    plugins: [
-      plugins.define({
-        __DEVELOPMENT__: stage === `develop` || stage === `develop-html`
-      })
-    ]
-  });
-};
+exports.onCreateWebpackConfig = require("./gatsby/create-webpack-config");
 
 /**
  * Add the MDX extensions as resolvable. This is how the page creator
@@ -140,6 +46,9 @@ exports.preprocessSource = async function preprocessSource(
   return null;
 };
 
+/**
+ * Required config for mdx to function
+ */
 exports.onCreateBabelConfig = ({ actions }) => {
   actions.setBabelPlugin({
     name: `@babel/plugin-proposal-object-rest-spread`
