+++ packages/gatsby-mdx/index.js
@@ -1,1 +1,5 @@
-// noop
+/**
+ * Welcome to gastby-mdx!
+ *
+ * Start reading in gatsby-node.js
+ */
